refactor: Modify JwtDto path with auth package
