+++ src/main/java/org/ikuzo/otboo/global/security/JwtLoginSuccessHandler.java
@@ -9,6 +9,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.ikuzo.otboo.domain.auth.dto.JwtDto;
 import org.ikuzo.otboo.global.exception.ErrorResponse;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.Authentication;
